Step4 (#3246)
* refactor : plus 연산 리턴 Money 객체 생성

* fix : 테스트를 위한 보너스번호 생성 수정

* feat : 수동티켓 추가. (수동과 자동 티켓 함께 생성)

* view : 수동티켓 추가로 입력부분 추가

* refactor : 상금 자료형 int to long

* feat : 로또 수동발급 티켓개수 검증

* view : 수동발급 입출력 추가

* refactor : 인수가 필요없는stream generate()로 변경

* refactor : 숫자배열을 받아 로또번호 리스트를 만들어 주는 test util class

* refactor : 구매 비용으로 티켓을 살 수 없는 경우 예외처리

* refactor : 테스트용 클래스 test package로 이동

+++ src/main/java/lotto/model/LottoResult.java
@@ -18,16 +18,17 @@ public class LottoResult {
     }
 
     public double getRateResult(int ticketCount) {
-        int totalReward = getTotalReward();
+        long totalReward = getTotalReward();
         Money totalMoney = new Money(totalReward);
+
         return totalMoney.divide(new Money(ticketCount * Cost.ONE_LOTTO_TICKET_PRICE));
     }
 
-    public int getTotalReward() {
+    public long getTotalReward() {
         return lottoResult.entrySet()
                 .stream()
                 .filter(o -> o.getValue() > 0)
-                .mapToInt(o -> o.getKey().prize() * o.getValue())
+                .mapToLong(o -> o.getKey().prize() * o.getValue())
                 .sum();
     }
